Revert "Remove dbt libraries from script"
